feat: make station deletion hard instead of soft
- Changed station_delete view to hard delete stations
- Changed delete_station_api to hard delete instead of setting active=False
- Removed station_restore view and restore_station_api
- Removed restore station URLs from both urls.py and api_urls.py
- Save station name before deletion to preserve it in success message

+++ core/models/dry_station.py
@@ -20,7 +20,7 @@ class DryQuestion(TimestampMixin):
     question_text = models.TextField()
     stem_image = models.CharField(max_length=255, blank=True, default='')
 
-    points = models.IntegerField(default=1)
+    points = models.FloatField(default=1)
 
     # Essay specific
     instructions = models.TextField(blank=True, default='')
